+++ liquid/golden/last_filter.py
@@ -8,4 +8,33 @@ cases = [
         template=r"{{ (1..3) | last }}",
         expect="3",
     ),
+    Case(
+        description="array of strings",
+        template=r"{{ arr | last }}",
+        expect="b",
+        globals={"arr": ["a", "b"]},
+    ),
+    Case(
+        description="array of things",
+        template=r"{{ arr | last }}",
+        expect="{}",
+        globals={"arr": ["a", "b", 1, [], {}]},
+    ),
+    Case(
+        description="empty array",
+        template=r"{{ arr | last }}",
+        expect="",
+        globals={"arr": []},
+    ),
+    Case(
+        description="left value not an array",
+        template=r"{{ arr | last }}",
+        expect="",
+        globals={"arr": 12},
+    ),
+    Case(
+        description="left value is undefined",
+        template=r"{{ nosuchthing | last }}",
+        expect="",
+    ),
 ]