+++ tests/test_render.py
@@ -3,18 +3,22 @@
 import asyncio
 import unittest
 
+from typing import List
+
 from liquid.environment import Environment
 from liquid.template import AwareBoundTemplate
 from liquid.mode import Mode
 from liquid.loaders import DictLoader
 
 from liquid import golden
+from liquid.exceptions import Error
+from liquid.golden.case import Case
 
 
 class RenderTestCase(unittest.TestCase):
     def _test(
         self,
-        test_cases,
+        test_cases: List[Case],
         template_class=AwareBoundTemplate,
         tolerance=Mode.STRICT,
     ):
@@ -24,7 +28,7 @@ class RenderTestCase(unittest.TestCase):
 
     def _test_sync(
         self,
-        test_cases,
+        test_cases: List[Case],
         template_class=AwareBoundTemplate,
         tolerance=Mode.STRICT,
     ):
@@ -36,15 +40,19 @@ class RenderTestCase(unittest.TestCase):
             )
             env.template_class = template_class
 
-            template = env.from_string(case.template, globals=case.globals)
-
-            with self.subTest(msg=case.description):
-                result = template.render()
-                self.assertEqual(result, case.expect)
+            with self.subTest(msg=case.description, mode=tolerance):
+                if case.error:
+                    with self.assertRaises(Error):
+                        template = env.from_string(case.template, globals=case.globals)
+                        result = template.render()
+                else:
+                    template = env.from_string(case.template, globals=case.globals)
+                    result = template.render()
+                    self.assertEqual(result, case.expect)
 
     def _test_async(
         self,
-        test_cases,
+        test_cases: List[Case],
         template_class=AwareBoundTemplate,
         tolerance=Mode.STRICT,
     ):
@@ -60,11 +68,15 @@ class RenderTestCase(unittest.TestCase):
             )
             env.template_class = template_class
 
-            template = env.from_string(case.template, globals=case.globals)
-
-            with self.subTest(msg=case.description, asynchronous=True):
-                result = asyncio.run(coro(template))
-                self.assertEqual(result, case.expect)
+            with self.subTest(msg=case.description, asynchronous=True, mode=tolerance):
+                if case.error:
+                    with self.assertRaises(Error):
+                        template = env.from_string(case.template, globals=case.globals)
+                        result = asyncio.run(coro(template))
+                else:
+                    template = env.from_string(case.template, globals=case.globals)
+                    result = asyncio.run(coro(template))
+                    self.assertEqual(result, case.expect)
 
 
 class RenderTagsTestCase(RenderTestCase):
@@ -138,7 +150,7 @@ class RenderTagsTestCase(RenderTestCase):
 
     def test_illegal(self):
         """Test that we can render an `illegal` token in LAX mode."""
-        self._test(golden.illegal_tag.cases, tolerance=Mode.LAX)
+        self._test(golden.illegal_tag.cases)
 
     def test_include_tag(self):
         """Test that we can render `include` tags."""
@@ -169,3 +181,31 @@ class RenderFiltersTestCase(RenderTestCase):
     def test_render_last_filter(self):
         """Test that we render the `last` filter correctly."""
         self._test(golden.last_filter.cases)
+
+    def test_render_map_filter(self):
+        """Test that we render the `map` filter correctly."""
+        self._test(golden.map_filter.cases)
+
+    def test_render_reverse_filter(self):
+        """Test that we render the `reverse` filter correctly."""
+        self._test(golden.reverse_filter.cases)
+
+    def test_render_sort_filter(self):
+        """Test that we render the `sort` filter correctly."""
+        self._test(golden.sort_filter.cases)
+
+    def test_render_sort_natural_filter(self):
+        """Test that we render the `sort_natural` filter correctly."""
+        self._test(golden.sort_natural_filter.cases)
+
+    def test_render_where_filter(self):
+        """Test that we render the `where` filter correctly."""
+        self._test(golden.where_filter.cases)
+
+    def test_render_uniq_filter(self):
+        """Test that we render the `uniq` filter correctly."""
+        self._test(golden.uniq_filter.cases)
+
+    def test_render_compact_filter(self):
+        """Test that we render the `compact` filter correctly."""
+        self._test(golden.compact_filter.cases)